etl/requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.3
pandas~=2.2.3
PyPDF2~=3.0.1
mysql-connector-python~=9.2.0